Response\HeadersTest: split `flatten()` specific tests to dedicated test class
