+++ tests/Response/Headers/HeadersTest.php
@@ -283,62 +283,4 @@ final class HeadersTest extends TestCase {
 			}
 		}
 	}
-
-	/**
-	 * Tests flattening of data.
-	 *
-	 * @covers ::flatten
-	 *
-	 * @dataProvider dataFlatten
-	 *
-	 * @param string|array $input    Value to flatten.
-	 * @param string       $expected Expected output value.
-	 *
-	 * @return void
-	 */
-	public function testFlatten($input, $expected) {
-		$headers = new Headers();
-		$this->assertSame($expected, $headers->flatten($input));
-	}
-
-	/**
-	 * Data Provider.
-	 *
-	 * @return array
-	 */
-	public function dataFlatten() {
-		return [
-			'string'            => ['text', 'text'],
-			'empty array'       => [[], ''],
-			'array with values' => [['text', 10, 'more text'], 'text,10,more text'],
-		];
-	}
-
-	/**
-	 * Tests receiving an exception when an invalid value is passed to flatten().
-	 *
-	 * @covers ::flatten
-	 *
-	 * @dataProvider dataFlattenInvalidValue
-	 *
-	 * @param mixed $input Value to flatten.
-	 *
-	 * @return void
-	 */
-	public function testFlattenInvalidValue($input) {
-		$this->expectException(InvalidArgument::class);
-		$this->expectExceptionMessage('Argument #1 ($value) must be of type string|array');
-
-		$headers = new Headers();
-		$headers->flatten($input);
-	}
-
-	/**
-	 * Data Provider.
-	 *
-	 * @return array
-	 */
-	public function dataFlattenInvalidValue() {
-		return TypeProviderHelper::getAllExcept(TypeProviderHelper::GROUP_STRING, TypeProviderHelper::GROUP_ARRAY);
-	}
 }
